Update default certificate digest and fix for Python 3
MD5 is not accepted anymore by even intermediate OpenSSL versions. SHA256 is a current standard for a longer while. Similarly DSA not seen as insecure nowadays and we didn't use it, hence remove it from the library.

Python 3 does not support self-expanding tuples as function arguments anymore. Also those are unnecessary here. Furthermore cert and key writes need to be done explicitly in bytes type, to match the output buffers given by pyOpenSSL functions.

Do not overwrite built-in "type" with createKeyPair() function.

Signed-off-by: MichaIng <micha@dietpi.com>